Added docs and test script for building in Conda
Use a local Conda environment since users don't need sudo
priviliedges to do this on a compute cluster AND one can
run these exact commands on the GPU nodes that will be
running the code in production, later.

The test_tensorflow.py is helpful for benchmarking CPU
and GPU usage in the Conda environment or in an official
TF container image.  Initial tests suggest they run at
the same speed.